Add namespace to framework chart
This will create the namespace when the framework addon is enabled, and
remove it when it is disabled / when the ManifestWork is removed.

Previously, the namespace was specially created by the status sync
controller, and removed via a special job. Changing the way it is
managed will help us adjust some permissions we don't otherwise need.

Refs:
 - https://github.com/stolostron/backlog/issues/20826

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>