Adjust tests to correctly deploy on the hub
Adding the cluster namespace to the framework chart meant that there
is another difference between the resources created when deployed to the
hub cluster. In particular, if the default framework addon is deployed
to the hub, then problems will come up when it is removed, and the
cluster namespace is incorrectly deleted.

Now tests will use a different framework CR when deploying to the hub,
and some tests are adjusted to only apply to the hub/managed clusters.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>