Add startup probes to cert and iam controllers
The startupProbes will help prevent a slow API discovery from sending
the containers into CrashLoopBackOff. On clusters where startupProbes
might not be available (pre k8s 1.20), we have to use an equivalent
initialDelaySeconds parameter.

In order to restart the pods more quickly on the managed clusters,
especially during testing, this also updates the deployment strategy
to "Recreate". Thus it will not wait for the previous deployment to
become ready before deleting it, which would take a minimum of 5
minutes on older k8s clusters.

Refs:
 - https://github.com/stolostron/backlog/issues/23874

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>