Add Prometheus monitoring to the config-policy-controller
If on OpenShift, the kube-rbac-proxy is used to protect the /metrics
endpoint. The ServiceMonitor is created in the openshift-monitoring
namespace so that the metrics are scraped automatically in the OpenShift
Prometheus instance over HTTPS.

If not on OpenShift, the ServiceMonitor is created in the addon
namespace and a Prometheus instance must be configured to look for
ServiceMonitors in the addon namespace. The metrics are served over HTTP
in this case.

Note that the created Role and RoleBinding are to give the Prometheus
operator access to the addon namespace during its discovery of what to
scrape.

Related:
https://github.com/stolostron/backlog/issues/23547

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 318c4d17e12cc6eac4dfa2a3ad4d844986fa5cb1)