Disable leader election when the replicas are set to 1
The addons have the deployment strategy of "Recreate", so leader
election is not applicable when the replicas are set to 1. In this
case, leader election is disabled to reduce resource utilization on
the Kubernetes API server.

Relates:
https://github.com/stolostron/backlog/issues/26711
https://issues.redhat.com/browse/ACM-1867

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 0225a873debb33d7eb104633e5d47ccc26a6d7bd)