Disable leader election when the replicas are set to 1
The addons have the deployment strategy of "Recreate", so leader
election is not applicable when the replicas are set to 1. In this
case, leader election is disabled to reduce resource utilization on
the Kubernetes API server.

Relates:
https://github.com/stolostron/backlog/issues/26711
https://issues.redhat.com/browse/ACM-1867

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit ad077afb1fe1dc2287b622bbd8a4f7b3fbc740e4)