Double the framework-addon's default resources
When the status-sync controller starts, it looks through all events on
the cluster that involves a policy. This may be a huge number of events,
and in some instances it has exceeded the memory limit, resulting in
Kubernetes killing the process.

Previously, the spec-sync, status-sync, and template-sync controllers
were in separate containers, each with their own memory requests and
limits. When they were combined into one container, those values  per
container were not changed. After this change's increase, the requests
and limits are still lower for the pod overall compared to when the
controllers were combined.

Refs:
 - https://issues.redhat.com/browse/ACM-3879

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit e9acb64916b54aeeace7e9ed1bef080b891aaafb)