Add client QPS and Burst customization to cfg-pol
This allows the new `--client-max-qps` and `--client-burst` arguments
on the configuration-policy-controller to be modified by the user. If
those are not explicitly configured, their values will scale with the
configured evaluation concurrency.

As part of this change, more values will be omitted when empty, which
allows for the default values from the `values.yaml` to be used, and
allows for other layers of configuration (eg AddOnDeploymentConfig)
to be effective, as opposed to the deafult in `getValues` always
being used.

Refs:
 - https://issues.redhat.com/browse/ACM-4297

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit ea6aaacc4e7ec350eb1153614951041a8cda0448)