Stop deleting Policy CRD on hub
The previous attempt to address this issue actually caused the issue in
upgrade situations. The new approach ensures that the CRD will not be
deleted on the hub when the addon is removed.

There is still a potential conflict because this CRD will be defined on
the hub by this controller, and another installer. In a future release,
after this "deletion-orphan" annotation has been added, it might be
possible to stop defining the CRD on the hub through this controller.

Refs:
 - https://issues.redhat.com/browse/ACM-4917

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 26036639ec5c9cd5841efc3e0ace0b9984f266d0)