Enhance crd-sync to add labels and annotations
The previous update to the Policy CRD which added the "deletion-orphan"
annotation would not be persisted through the "crd-sync.sh" script. Now,
that script adds the annotation.

The script also adds the "policy-type" annotation to the v1beta1 CRD for
ConfigurationPolicy, since that was missing. It is easier to add it here
rather than update the kustomization in the other repo, since it never
creates the v1beta1 crd in that repo's scripts.

Refs:
 - https://issues.redhat.com/browse/ACM-5326

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 42101740ecc64b3f82e09ccb733a43242a881ea4)