Set the uninstallation annotation in the chart
In certain situations, if the uninstall procedure was interrupted with
a reinstallation, the uninstallation annotations could still be on the
managed cluster, in the wrong state. Now, the manifest work should also
help manage that annotation, ensuring it will be set to "false" when the
policy addons are not supposed to be being uninstalled.

Refs:
 - https://issues.redhat.com/browse/ACM-6941

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit c1e255e4bb58eec0f02462c4c27d9e8d638e958c)