Add support for explicitly enabling the spec sync controller
Normally, the spec sync is disabled if it's a self-managed hub. The
caveat is when a hub is imported by another hub, the spec sync should
still sync from the top-level hub.

Additionally, this allows the on-multicluster-hub annotation to be set
on the ManagedCluster object instead of the addon directly.

Relates:
https://issues.redhat.com/browse/ACM-7150

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 91f6c2aa14756d858183fd630ea80c5a1926c527)