Create the cluster namespace when a hub is imported by a global hub
When syncPoliciesOnMulticlusterHub is true, it's indicating that the hub
is imported by a global hub. The namespace must be created on the hub to
sync policies from the global hub. The namespace also needs to be
cleaned up when uninstalled, so the clean up pod is now enabled in this
scenario too.

Relates:
https://issues.redhat.com/browse/ACM-7655

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 21c01f09f46b48d119036da819a0af0a3278a52c)