Update the addon-framework to fix uninstalls
This allows updates to the config-policy-controller Deployment manifest
while the uninstall pod is running. In particular, this allows the
uninstall annotation to be set on the Deployment when the addon is
uninstalling and it allows for container image updates if an uninstall
is stuck.

Relates:
https://issues.redhat.com/browse/ACM-8947
https://github.com/open-cluster-management-io/addon-framework/commit/34f78ece0a09cd64bc66620e2bcc230113b3b60d

Signed-off-by: mprahl <mprahl@users.noreply.github.com>
(cherry picked from commit 3cebe88543e26272e8e2a67777e80f33346491a4)