Set OperatorPolicy default namespace option
This sets the command-line flag added by the companion PR
https://github.com/open-cluster-management-io/config-policy-controller/pull/204

This only sets that flag when the managed cluster is using OpenShift v4.
It should be noted that this flag is *not* currently configurable by the
user (it is not for example exposed through an annotation).

Refs:
 - https://issues.redhat.com/browse/ACM-9896

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 010070d5d134801cb3a257bf865f216318156b58)