Add the ocm-policies namespace to the uninstall
ConfigurationPolicies in that namespace should be checked for finalizers
so that the uninstall can complete. It should not be checked when the
addon is in hosted mode, since in that case the namespace might be used
by other clusters.

Refs:
 - https://issues.redhat.com/browse/ACM-14707

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 1137556a4ab5bc18cecf45d17a59d98f0a7ba8b9)