Add standalone-hub-templating addon
This addon provides a mechanism for the configuration-policy-controller
(including both ConfigurationPolicy and OperatorPolicy) to access
resources on the hub cluster in order to resolve "hub templates." The
addon itself is very small, it is just some base RBAC resources and a
secret on the managed cluster to help the user identify how to configure
additional permissions. Mainly, the config-policy-controller needs to be
configured to mount the new addon's hub kubeconfig and utilize it.

Refs:
 - https://issues.redhat.com/browse/ACM-16089

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 9d1a7372f38e89b7d42c766808960f8160b14555)