Use informers for MC and MCAO lookups
In a scale test, it was found that this controller was falling behind
when it needed to reconcile thousands of clusters. The main cause was
the new GET request when reconciling the config-policy-controller addon,
since it needs to check whether the governance-standalone-hub-templating
addon for that managed cluster exists. At heavy loads, this caused each
reconcile to take approximately 0.2 seconds because of rate-limiting.

That GET request, as well as a similar requests for ManagedClusters done
by multiple addons in hosted mode, will now use informers to use a local
cache of those resources. These "requests" are now basically instant.

Refs:
 - https://issues.redhat.com/browse/ACM-18029

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
(cherry picked from commit 9c0c8a282b5d478aaee5b8ca42200f452917aa7d)